+++ R/later.R
@@ -1,5 +1,4 @@
-#' @useDynLib later
-#' @import Rcpp
+#' @useDynLib later, .registration=TRUE
 #' @importFrom Rcpp evalCpp
 
 .onLoad <- function(...) {
@@ -14,6 +13,17 @@
 # get the corresponding loop handle. We use weak refs because we don't want
 # this registry to keep the loop objects alive.
 .loops <- new.env(parent = emptyenv())
+
+# Our own weakref functions are implemented (instead of using those from
+# `rlang`) to avoid loading `rlang` automatically upon package load, as this
+# causes additional overhead for packages which only link to `later`.
+new_weakref <- function(loop) {
+  .Call(`_later_new_weakref`, loop)
+}
+
+wref_key <- function(w) {
+  .Call(`_later_wref_key`, w)
+}
 
 #' Private event loops
 #'
@@ -94,7 +104,7 @@ create_loop <- function(parent = current_loop(), autorun = NULL) {
   lockBinding("id", loop)
 
   # Add a weak reference to the loop object in our registry.
-  .loops[[sprintf("%d", id)]] <- rlang::new_weakref(loop)
+  .loops[[as.character(id)]] <- new_weakref(loop)
 
   if (id != 0L) {
     # Inform the C++ layer that there are no more R references when the handle
@@ -103,7 +113,7 @@ create_loop <- function(parent = current_loop(), autorun = NULL) {
     # However, if the package is unloaded it can get GC'd, and we don't want the
     # destroy_loop() finalizer to give an error message about not being able to
     # destroy the global loop.
-    reg.finalizer(loop, notify_r_ref_deleted)
+    reg.finalizer(loop, notify_r_ref_deleted, onexit = TRUE)
   }
 
   loop
@@ -116,7 +126,7 @@ notify_r_ref_deleted <- function(loop) {
 
   res <- notifyRRefDeleted(loop$id)
   if (res) {
-    rm(list = sprintf("%d", loop$id), envir = .loops)
+    rm(list = as.character(loop$id), envir = .loops)
   }
   invisible(res)
 }
@@ -130,7 +140,7 @@ destroy_loop <- function(loop) {
 
   res <- deleteCallbackRegistry(loop$id)
   if (res) {
-    rm(list = sprintf("%d", loop$id), envir = .loops)
+    rm(list = as.character(loop$id), envir = .loops)
   }
   invisible(res)
 }
@@ -145,12 +155,12 @@ exists_loop <- function(loop) {
 #' @export
 current_loop <- function() {
   id <- getCurrentRegistryId()
-  loop_weakref <- .loops[[sprintf("%d", id)]]
+  loop_weakref <- .loops[[as.character(id)]]
   if (is.null(loop_weakref)) {
     stop("Current loop with id ", id, " not found.")
   }
 
-  loop <- rlang::wref_key(loop_weakref)
+  loop <- wref_key(loop_weakref)
   if (is.null(loop)) {
     stop("Current loop with id ", id, " not found.")
   }
@@ -249,8 +259,12 @@ print.event_loop <- function(x, ...) {
 #'
 #' @export
 later <- function(func, delay = 0, loop = current_loop()) {
-  f <- rlang::as_function(func)
-  id <- execLater(f, delay, loop$id)
+  # `rlang::as_function` is used conditionally so that `rlang` is not loaded
+  # until used, avoiding this overhead for packages only linking to `later`
+  if (!is.function(func)) {
+    func <- rlang::as_function(func)
+  }
+  id <- execLater(func, delay, loop$id)
 
   invisible(create_canceller(id, loop$id))
 }
@@ -263,6 +277,98 @@ create_canceller <- function(id, loop_id) {
   force(loop_id)
   function() {
     invisible(cancel(id, loop_id))
+  }
+}
+
+#' Executes a function when a file descriptor is ready
+#'
+#' Schedule an R function or formula to run after an indeterminate amount of
+#' time when file descriptors are ready for reading or writing, subject to an
+#' optional timeout.
+#'
+#' On the occasion the system-level `poll` (on Windows `WSAPoll`) returns an
+#' error, the callback will be made on a vector of all `NA`s. This is
+#' indistinguishable from a case where the `poll` succeeds but there are error
+#' conditions pending against each file descriptor.
+#'
+#' If no file descriptors are supplied, the callback is scheduled for immediate
+#' execution and made on the empty logical vector `logical(0)`.
+#'
+#' @param func A function that takes a single argument, a logical vector that
+#'   indicates which file descriptors are ready (a concatenation of `readfds`,
+#'   `writefds` and `exceptfds`). This may be all `FALSE` if the
+#'   `timeout` argument is non-`Inf`. File descriptors with error conditions
+#'   pending are represented as `NA`, as are invalid file descriptors such as
+#'   those already closed.
+#' @param readfds Integer vector of file descriptors, or Windows SOCKETs, to
+#'   monitor for being ready to read.
+#' @param writefds Integer vector of file descriptors, or Windows SOCKETs, to
+#'   monitor being ready to write.
+#' @param exceptfds Integer vector of file descriptors, or Windows SOCKETs, to
+#'   monitor for error conditions pending.
+#' @param timeout Number of seconds to wait before giving up, and calling `func`
+#'   with all `FALSE`. The default `Inf` implies waiting indefinitely.
+#'   Specifying `0` will check once without blocking, and supplying a negative
+#'   value defaults to a timeout of 1s.
+#' @param loop A handle to an event loop. Defaults to the currently-active loop.
+#'
+#' @inherit later return note
+#'
+#' @examplesIf requireNamespace("nanonext", quietly = TRUE)
+#' # create nanonext sockets
+#' s1 <- nanonext::socket(listen = "inproc://nano")
+#' s2 <- nanonext::socket(dial = "inproc://nano")
+#' fd1 <- nanonext::opt(s1, "recv-fd")
+#' fd2 <- nanonext::opt(s2, "recv-fd")
+#'
+#' # 1. timeout: prints FALSE, FALSE
+#' later_fd(print, c(fd1, fd2), timeout = 0.1)
+#' Sys.sleep(0.2)
+#' run_now()
+#'
+#' # 2. fd1 ready: prints TRUE, FALSE
+#' later_fd(print, c(fd1, fd2), timeout = 1)
+#' res <- nanonext::send(s2, "msg")
+#' Sys.sleep(0.1)
+#' run_now()
+#'
+#' # 3. both ready: prints TRUE, TRUE
+#' res <- nanonext::send(s1, "msg")
+#' later_fd(print, c(fd1, fd2), timeout = 1)
+#' Sys.sleep(0.1)
+#' run_now()
+#'
+#' # 4. fd2 ready: prints FALSE, TRUE
+#' res <- nanonext::recv(s1)
+#' later_fd(print, c(fd1, fd2), timeout = 1)
+#' Sys.sleep(0.1)
+#' run_now()
+#'
+#' # 5. fds invalid: prints NA, NA
+#' close(s2)
+#' close(s1)
+#' later_fd(print, c(fd1, fd2), timeout = 0)
+#' Sys.sleep(0.1)
+#' run_now()
+#'
+#' @export
+later_fd <- function(func, readfds = integer(), writefds = integer(), exceptfds = integer(),
+                     timeout = Inf, loop = current_loop()) {
+  if (!is.function(func)) {
+    func <- rlang::as_function(func)
+  }
+  xptr <- execLater_fd(func, readfds, writefds, exceptfds, timeout, loop$id)
+
+  invisible(create_fd_canceller(xptr))
+}
+
+# Returns a function that will cancel a callback with the given external
+# pointer. If the callback has already been executed or canceled, then the
+# function has no effect.
+create_fd_canceller <- function(xptr) {
+  force(xptr)
+  function() {
+    invisible(fd_cancel(xptr))
   }
 }