+++ src/callback_registry.h
@@ -56,7 +56,15 @@ public:
   StdFunctionCallback(Timestamp when, std::function<void (void)> func);
 
   void invoke() const {
+#ifdef RCPP_USING_UNWIND_PROTECT // See https://github.com/r-lib/later/issues/191
+    Rcpp::unwindProtect([this]() {
+      BEGIN_RCPP
+      func();
+      END_RCPP
+    });
+#else
     func();
+#endif
   }
 
   Rcpp::RObject rRepresentation() const;
@@ -107,6 +115,7 @@ private:
   // objects to be copied on the wrong thread, and even trigger an R GC event
   // on the wrong thread. https://github.com/r-lib/later/issues/39
   cbSet queue;
+  int fd_waits = 0;
   Mutex* mutex;
   ConditionVariable* condvar;
 
@@ -134,7 +143,7 @@ public:
   // Use this to determine the next time we need to pump events.
   Optional<Timestamp> nextTimestamp(bool recursive = true) const;
 
-  // Is the registry completely empty?
+  // Is the registry completely empty? (including later_fd waits)
   bool empty() const;
 
   // Is anything ready to execute?
@@ -148,6 +157,10 @@ public:
 
   // Return a List of items in the queue.
   Rcpp::List list() const;
+
+  // Increment and decrement the number of active later_fd waits
+  void fd_waits_incr();
+  void fd_waits_decr();
 
   // References to parent and children registries. These are used for
   // automatically running child loops. They should only be accessed and